requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
Markdown==3.10
MarkupSafe==3.0.3
Pygments==2.17.2
python-slugify==8.0.4
text-unidecode==1.3
Werkzeug==3.1.3
gunicorn==23.0.0
